+++ b4n/kube/resources/data/mod.rs
@@ -30,7 +30,9 @@ pub mod persistent_volume;
 pub mod persistent_volume_claim;
 pub mod pod;
 pub mod pod_metrics;
+pub mod priority_class;
 pub mod replica_set;
+pub mod role_binding;
 pub mod secret;
 pub mod service;
 pub mod service_account;
@@ -70,7 +72,9 @@ pub fn get_resource_data(
         "PersistentVolumeClaim" => persistent_volume_claim::data(object),
         "Pod" => pod::data(object, stats),
         "PodMetrics" if group == "metrics.k8s.io" => pod_metrics::data(object),
+        "PriorityClass" => priority_class::data(object),
         "ReplicaSet" => replica_set::data(object),
+        "RoleBinding" => role_binding::data(object),
         "Secret" => secret::data(object),
         "Service" => service::data(object),
         "ServiceAccount" => service_account::data(object),
@@ -107,7 +111,9 @@ pub fn get_header_data(kind: &str, group: &str, crd: Option<&CrdColumns>, has_me
         "PersistentVolumeClaim" => persistent_volume_claim::header(),
         "Pod" => pod::header(has_metrics),
         "PodMetrics" if group == "metrics.k8s.io" => pod_metrics::header(),
+        "PriorityClass" => priority_class::header(),
         "ReplicaSet" => replica_set::header(),
+        "RoleBinding" => role_binding::header(),
         "Secret" => secret::header(),
         "Service" => service::header(),
         "ServiceAccount" => service_account::header(),